Alert

+++ components/FavoritesComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
@@ -15,8 +15,8 @@ const mapStateToProps = state => {
 };
 
 const mapDispatchToProps = {
-        deleteFavorite: campsiteId => (deleteFavorite(campsiteId))
-    };
+    deleteFavorite: campsiteId => (deleteFavorite(campsiteId))
+};
 
 class Favorites extends Component {
 
@@ -25,13 +25,32 @@ class Favorites extends Component {
     }
 
     render() {
+
         const { navigate } = this.props.navigation;
-        const renderFavoriteItem = ({item}) => {
+
+        const renderFavoriteItem = ({ item }) => {
             const rightButton = [
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite campsite ' + item.name + '?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ];
 
@@ -40,8 +59,8 @@ class Favorites extends Component {
                     <ListItem
                         title={item.name}
                         subtitle={item.description}
-                        leftAvatar={{source: {uri: baseUrl + item.image}}}
-                        onPress={() => navigate('CampsiteInfo', {campsiteId: item.id})}
+                        leftAvatar={{ source: { uri: baseUrl + item.image } }}
+                        onPress={() => navigate('CampsiteInfo', { campsiteId: item.id })}
                     />
                 </Swipeout>
             );
@@ -57,6 +76,7 @@ class Favorites extends Component {
                 </View>
             );
         }
+        
         return (
             <FlatList
                 data={this.props.campsites.campsites.filter(
